Design editProfile page with avatar on the left and other content on the right
