+++ data/frontend/src/Base.tsx
@@ -8,14 +8,13 @@ import Navigation from "./components/Navigation/Navigation";
 import NavDrawer from "./components/Navigation/NavDrawer";
 import Game from "./components/Game/Game";
 import Options from "./components/Global/Options";
-import Profile from "./components/Global/Profile";
+import Profile from "./components/Profile/Profile";
 import Home from "./components/Home/Home";
 import Chat from "./components/Chat/Chat";
 import Achievements from "./components/Achievements/Achievements";
-import Modification from "./components/Global/Modification";
+import EditProfile from "./components/Profile/EditProfile";
 import Friendlist from "./components/Friendlist/Friendlist";
 import webSocketManager from "./webSocket";
-import TwoFactorLogin from "./components/Login/TwoFactorLogin";
 
 export default function Base() {
   const user = useSelector((state: any) => state.user);
@@ -67,7 +66,7 @@ export default function Base() {
           element={<Achievements />}
         ></Route>
         <Route path="/friendlist" element={<Friendlist />}></Route>
-        <Route path="/edit" element={<Modification />}></Route>
+        <Route path="/editProfile" element={<EditProfile />}></Route>
       </Routes>
       {location.pathname !== "/2fa" ? <Chat /> : null}
     </div>
